Make reblog non-tree contributed content handling more robust. Fixes #644
Sometimes, there will be reblog "contributed content" (from the most
recent user) without there being a reblog tree. an example for this is
the first reblog of an ask with an answer, a reblog with text content
of a media post that didn't have text content before, or (I think)
a text post where previous reblogs were removed. This makes their
handeling more robust.