Add messaging_tweaks extension
This extension adds a bunch of small tweaks to the chat
system that tumblr has released recently. This version
features hiding the chat bubble entirely, allow the chat tab
to be hidden, and make the user's chat bubble look
differently for easy distinguishing while chatting.
You can also edit the background color of the chat window