Add Messaging Tweaks for Notifications
Many users asked for better notifications for the chat messages, since
it's quite easy to accidentally miss something new in the chat,
especially with the minimise chat tweak enabled and the chat hidden.
This update adds 3 types of notifications to the chat, sound, desktop
and also changing of tab title.

Additionally it adds a small tweak @nightpool wanted, having the
ability to put the chat partner's messages on the right as well