Fix broken :+1: emoji
The file of the :+1: emoji is called plus1.png instead of +1.png, which
the messaging tweaks has not been taking into account. This is fixed by
simply setting the emoji_url to the proper url instead of the generated
one. Furthermore, the text to be replaced had to be adjusted as well,
since `+` is a special character in RegEx and does not automatically get
escaped by `new RegExp`