Centralize Pokedex.JSON URL
Before, the Pokédex.json URL was in the code twice, but got changed due
to a mistake in the gist. @ThePsionic only changed 1 of the 2 urls to
point to the new gist though, which in future gist revisions could
have caused problems. Thus, centralizing and only having the URL once
is a better option here.