Add video ads to anti_capitalism
It seems that Tumblr implemented another ad-container for video ads. The
container with the class "video-ad-container" should now be successfully
blocked by anti-capitalism as well. This is untested though, since as of
now, no video ad has come up for me. The class comes from a helpful user
in the live support chat.