Make users opt-in to ER, for better UX.
Currently, I see a lot of posts that are "inadvertantly broken". It's really
useful to have ER installed for sometimes, but most people don't need it for
mosts posts. So instead, we can have everybody install it, and just have
people opt-in per post.

I'm guessing this will lead to less broken posts, while no degredation in UX
for people who really need and love ER.