Add tweak to allow much wider post sources
Tumblr sets post sources to have a max-width of 100px, which is very low. In particular there are blog usernames which are longer than this and get truncated with `...`. This eliminates that issue while still truncating very long source names.

I'm fairly certain that the reason for the 100px setting is because post tags used to start on the same line as sources, but that was changed when they introduced multi-line tags.

400px is somewhat arbitrary; it avoids truncation on anything reasonable with ~75px of slack space left over in English. Should be sufficient for any localized variant of "Source" and minor Tumblr changes.