User Menus+: fix identifying blog links (#1331)
User Menus+ responds to links to blogs, populating its internal
information with the popover that displays when you mouseover a link to
a blog. However, it was not identifying all such links, leading to
strange bugs where other extensions such as Profiler would display
information on the blog undefined.tumblr.com.

This introduces a more flexible method for identifying links to blogs
using a regular expression. User Menus+ now works correctly on links in
reblog headers.