Activity Feed Slim Tweak (#1326)
* Activity Feed Slim Tweak

Adds a tweak to re-slim the activity feed back to normal
Now not enabled by default
Tried to add readability
Added the Experimental tag in case imminent feature re-introductions break this or make it ugly

* Retags fix

Made retags less ugly when this is used

* Split into 2 tweaks

Highlighting mutuals and slimming the activity feed split into 2 tweaks

* Removed mutual highlighting

Now that default tumblr seems to highlight mutuals' notes again, I've removed the tweak for it. I'll happily move the remaining re-slimming tweak to Activity+ if that's tidier, and/or I could probably make this effect the Activity Button on the dashboard.

* Added updated Activity+ support

Moved the Activity+ timestamp back to where it was before.

* Defeated a tumblr update

* Prioritised current Retags behaviour over old look

Retags seems to insert tags after the Activity+ timestamp rather than before, as was the case before the tumblr update. The goal of this tweak was to give the activity feed its exact old look back but the way Retags and Activity+ timestamps work now, that is impossible, so this is the next best thing.