Changing the way likes are requested (#1436)
* Changing the way likes are requested

This fixes some issues with Mass Unliker only getting few posts... Hopefully. I couldn't reproduce the problem in the first place so I'm kinda writing blind. It does work though, I'm just not sure if it fixes the problem.

* fixed whitespace

* Style and whitespace changes

Speaking of, new lines are somewhat inconsistent and I'd like to fix it, but that's another PR.

* I missed one 

I hate whitespace

* I'm screaming eternally at whitespace