ReplyViewer: restore links to commenters posts
Fixed Issue #1118 using /svc/notes API as used in tagviewer. For reblogs will grab added text from notes if present but check post directly if not. Returned comment (if there is one) is in HTML so some stripping required. Bonus: shows you if a reply consisted of just an image (e.g. a gif) which notes and previous replyviewer don't.